requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
joblib>=1.4.0
fairlearn>=0.10.0
reportlab>=4.2.0
jinja2>=3.1.0
google-generativeai>=0.8.0
google-cloud-firestore>=2.16.0
python-dotenv>=1.0.0
pytest>=8.2.0
httpx>=0.27.0
